viewer: MODULARIZE the wasm module and widen the digest to int64
Two failures the emsdk build found that a native build cannot:

- `int` is 32 bits under wasm32, so writing an FNV-1a u32 keyframe digest
  through `int()` raised RangeDefect and every replay was rejected. The
  digest now travels as int64 on both sides of the JSON.
- the non-MODULARIZE emscripten output declares its own `var Module`, so the
  Worker's pre-set globals (locateFile, onRuntimeInitialized) were silently
  discarded and the .data package never resolved. The module is now a factory
  (RaidReplayModule), which is also what bullwhip ships.

Also: raid_load_replay catches Exception rather than CatchableError, so a
wasm-only Defect surfaces as a message in the shell instead of a silent zero
with an empty error string - which is how the RangeDefect was found.

+++ tests/test_determinism.nim
@@ -97,9 +97,9 @@ proc testNoFloatingPointInTheStepPath() =
         break
   done("no floating point anywhere in the step path")
 
-proc digestsOf(world: Sim): seq[int] =
+proc digestsOf(world: Sim): seq[int64] =
   for frame in world.keyframes:
-    result.add(int(frame.digest))
+    result.add(int64(frame.digest))
 
 proc testSameSeedSameDigests() =
   let config = testConfig()
@@ -151,7 +151,7 @@ proc testGoldenFixture() =
   let world = runScripted(config, skStalwart)
   var got = newJArray()
   for frame in world.keyframes:
-    got.add(%int(frame.digest))
+    got.add(%int64(frame.digest))
   let golden = parseJson(repoFile("tests/fixtures/golden_digests.json"))
   checkEq(golden{"game_version"}.getStr(), GameVersion,
     "the fixture was recorded against this GameVersion")
